[REFACTOR]: Provider 방식을 사용한 Flutter 전용 디자인 패턴
