+++ lib/pages/loading_page/loading_view.dart
@@ -1,0 +1,30 @@
+import 'package:flutter/material.dart';
+
+class LoadingViewModel with ChangeNotifier {
+  final PageController pageController = PageController();
+  int _currentPage = 0;
+
+  int get currentPage => _currentPage;
+
+  void changePage(int index) {
+    _currentPage = index;
+    notifyListeners();
+  }
+
+  void nextPage() {
+    if (_currentPage < 2) {
+      pageController.nextPage(
+        duration: const Duration(milliseconds: 300),
+        curve: Curves.easeInOut,
+      );
+      _currentPage++;
+      notifyListeners();
+    }
+  }
+
+  @override
+  void dispose() {
+    pageController.dispose();
+    super.dispose();
+  }
+}
