+++ lib/pages/loading_page/page_controller.dart
@@ -1,61 +1,42 @@
-import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
-import 'loading_page1.dart';
-import 'loading_page2.dart';
+import 'package:provider/provider.dart';
+import 'loading_page1/loading_page1.dart';
+import 'loading_page2/loading_page2.dart';
+import 'loading_view.dart';
 import 'loading_page3.dart';
-import 'package:smooth_page_indicator/smooth_page_indicator.dart';
 
-class LoginPageController extends StatefulWidget {
+class LoginPageController extends StatelessWidget {
   final String? kakaoNativeAppKey;
   final String? kakaoJavaScriptAppKey;
-  const LoginPageController({Key? key, required this.kakaoNativeAppKey, required this.kakaoJavaScriptAppKey}) : super(key: key);
 
-  @override
-  State<LoginPageController> createState() => _LoginPageControllerState();
-}
-
-class _LoginPageControllerState extends State<LoginPageController> {
-  final PageController _controller = PageController();
+  const LoginPageController({
+    super.key,
+    required this.kakaoNativeAppKey,
+    required this.kakaoJavaScriptAppKey,
+  });
 
   @override
   Widget build(BuildContext context) {
-    final height = MediaQuery.of(context).size.height;
-    double width = MediaQuery.of(context).size.width;
-    if (kIsWeb) {
-      width = 430;
-    }
+    final viewModel = Provider.of<LoadingViewModel>(context);
+
     return Scaffold(
       backgroundColor: Colors.white,
       body: Column(
         children: [
           Expanded(
             child: PageView(
-              controller: _controller,
-              scrollDirection: Axis.horizontal,
-              physics: const ClampingScrollPhysics(),
+              controller: viewModel.pageController,
+              onPageChanged: viewModel.changePage,
               children: [
                 const LoadingPage1(),
                 const LoadingPage2(),
                 LoadingPage3(
-                  kakaoNativeAppKey: widget.kakaoNativeAppKey,
-                  kakaoJavaScriptAppKey: widget.kakaoJavaScriptAppKey,
+                  kakaoNativeAppKey: kakaoNativeAppKey,
+                  kakaoJavaScriptAppKey: kakaoJavaScriptAppKey,
                 ),
               ],
             ),
           ),
-          Padding(
-            padding: EdgeInsets.only(bottom: height * 0.029),
-            child: SmoothPageIndicator(
-              controller: _controller,
-              count: 3,
-              effect: WormEffect(
-                dotHeight: height * 0.012,
-                dotWidth: width * 0.026,
-                activeDotColor: Colors.black,
-                dotColor: Colors.grey.shade300,
-              ),
-            ),
-          )
         ],
       ),
     );
